python-backend/requirements.txt
openai-agents
pydantic
fastapi
uvicorn
python-dotenv

# Voice Integration Dependencies
elevenlabs==0.2.26
pydub==0.25.1
speechrecognition==3.10.0

# Web Navigation Dependencies  
playwright==1.40.0
websockets==12.0

# Audio Processing (pyaudio has macOS compatibility issues, using browser-based audio instead)
# pyaudio==0.2.11
